+++ waspc/cli/src/Wasp/Cli/Command/Call.hs
@@ -31,7 +31,7 @@ data NewArgs = NewArgs
 
 data TestArgs = TestClient !Arguments | TestServer !Arguments deriving (Show, Eq)
 
-data StartArg = StartDb | StartNormal deriving (Show, Eq)
+data StartArg = StartDb | StartApp deriving (Show, Eq)
 
 data WaspLSArgs = WaspLSArgs
   { wslLogFile :: !(Maybe FilePath),